Reorganise raw data typemap,  so typemaps folder contain only common part. Improve document.

+++ Examples/test-suite/go/li_cdata_cpp_runme.go
@@ -6,7 +6,7 @@ func main() {
 	s := []byte("ABC\x00abc")
 	m := Malloc(256)
 	Memmove(m, s)
-	ss := Cdata(m, 7)
+	ss := Cdata(m, int64(7))
 	if string(ss) != "ABC\x00abc" {
 		panic("failed")
 	}
